+++ app/pages/register.php
@@ -3,17 +3,15 @@
 // registration is allowed, go on
 if ($config['registration_enabled'] === true) {
 
-    require_once '../app/classes/database.php';
     require '../app/classes/user.php';
     unset($error);
 
     try {
 
         // connect to database
-        require '../app/helpers/database.php';
-        $db = connectDB($config);
+        $dbWeb = connectDB($config);
 
-        $user = new User($db);
+        $user = new User($dbWeb);
 
         if ( $_SERVER['REQUEST_METHOD'] == 'POST' ) {
             $username = $_POST['username'];